+++ resources/libraries/get_orgs.rb
@@ -9,7 +9,11 @@ module RbManager
 
         organizations << m
       end
-
+      organizations.each do |org|
+        if org.override_attributes && org.override_attributes['redborder'] && org.override_attributes['redborder']['megabytes_limit'].is_a?(String) && org.override_attributes['redborder']['megabytes_limit'].strip.empty?
+          org.override_attributes['redborder']['megabytes_limit'] = nil
+        end
+      end
       organizations
     end
   end