+++ resources/recipes/configure.rb
@@ -278,7 +278,6 @@ end
 
 memcached_config 'Configure Memcached' do
   if manager_services['memcached']
-    memory node['redborder']['memory_services']['memcached']['memory']
     ipaddress node['ipaddress_sync']
     action [:add, :register]
   else
@@ -409,6 +408,8 @@ http2k_config 'Configure Http2k' do
     ips_nodes node.run_state['sensors_info']['ips-sensor']
     ipsg_nodes node.run_state['sensors_info']['ipsg-sensor']
     ipscp_nodes node.run_state['sensors_info']['ipscp-sensor']
+    intrusion_nodes node.run_state['sensors_info']['intrusion-sensor']
+    intrusioncp_nodes node.run_state['sensors_info']['intrusioncp-sensor']
     organizations node.run_state['organizations']
     locations_list node['redborder']['locations']
     action [:add, :register]
@@ -648,17 +649,21 @@ end
 # Determine external
 begin
   external_services = data_bag_item('rBglobal', 'external_services')
-rescue
-  external_services = {}
+rescue => e
+  Chef::Log.warn("Failed to load external_services data bag: #{e.message}")
+  external_services = nil
 end
 
 postgresql_config 'Configure postgresql' do
-  if manager_services['postgresql'] && external_services['postgresql'] == 'onpremise'
+  if manager_services['postgresql'] && external_services&.dig('postgresql') == 'onpremise'
     cdomain node['redborder']['cdomain']
     ipaddress node['ipaddress_sync']
     action [:add, :register]
-  else
+  elsif !external_services.nil?
     action [:remove, :deregister]
+  else
+    Chef::Log.warn('Skipped PostgreSQL removal/deregistration due to missing external_services data')
+    action :nothing
   end
 end
 
@@ -683,11 +688,14 @@ minio_config 'Configure S3 (minio)' do
   managers_with_minio node['redborder']['managers_per_services']['s3']
   access_key_id s3_secrets['s3_access_key_id']
   secret_key_id s3_secrets['s3_secret_key_id']
-  if manager_services['s3'] && (external_services['s3'] == 'onpremise')
+  if manager_services['s3'] && external_services&.dig('s3') == 'onpremise'
     ipaddress node['ipaddress_sync']
     action [:add_mcli, :add, :register]
-  else
+  elsif !external_services.nil?
     action [:add_mcli, :remove, :deregister]
+  else
+    Chef::Log.warn('Skipped MinIO removal/deregistration due to missing external_services data')
+    action :nothing
   end
 end
 
@@ -727,7 +735,7 @@ end
 
 # Configure Nginx s3 onpremise nodes for now..
 minio_config 'Configure Nginx S3 (minio)' do
-  if manager_services['s3'] && (external_services['s3'] == 'onpremise')
+  if manager_services['s3'] && external_services&.dig('s3') == 'onpremise'
     s3_hosts node['redborder']['s3']['s3_hosts']
     action [:add_s3_conf_nginx]
   elsif !manager_services['s3'] && manager_services['nginx']