+++ CHANGELOG.md
@@ -1,6 +1,28 @@
 cookbook-rb-manager CHANGELOG
 ===============
 
+## 5.1.0
+
+  - Miguel Alvarez
+    - [a93b7a5] Add intrusion & intrusioncp sensors
+    - [b512c3b] add intrusioncp to logstash pipelines
+    - [64d856c] Add intrusioncp sensor type
+
+## 5.0.3
+
+  - Rafael Gomez
+    - [91557b9] Improve error handling for external services in PostgreSQL and MinIO configurations
+
+## 5.0.2
+
+  - jnavarrorb
+    - [a4fa7e3] Fix empty string
+
+## 5.0.1
+
+  - Miguel Negrón
+    - [05cccc8] New memory distribution
+
 ## 5.0.0
 
   - Miguel Álvarez